refactor(auth): extract processInviteToken/claimConversations to plain modules
'use server' modules expose every export as a remote action, so testable
internals must live in plain modules. user/functions.ts + chat/functions.ts
now delegate to src/auth/invites.ts + src/chat/claims.ts.
